requirements/base.txt
django==5.2
djangorestframework==3.16.0
python-decouple==3.8
pillow==11.2.1
drf-spectacular==0.28.0
psycopg2-binary==2.9.10
drf-nested-routers==0.94.1
djangorestframework-simplejwt==5.4.0
django-treebeard==4.7.1
celery==5.5
redis==5.3.0
python-ulid==3.0.0
django-filter==25.1
httpx==0.28.1
gunicorn==23.0.0